+++ internal/examples/thrift-hello/hello/echo/helloserver/server.go
@@ -29,6 +29,7 @@ import (
 	transport "go.uber.org/yarpc/api/transport"
 	thrift "go.uber.org/yarpc/encoding/thrift"
 	echo "go.uber.org/yarpc/internal/examples/thrift-hello/hello/echo"
+	yarpcerrors "go.uber.org/yarpc/yarpcerrors"
 )
 
 // Interface is the server-side interface for the Hello service.
@@ -73,7 +74,8 @@ type handler struct{ impl Interface }
 func (h handler) Echo(ctx context.Context, body wire.Value) (thrift.Response, error) {
 	var args echo.Hello_Echo_Args
 	if err := args.FromWire(body); err != nil {
-		return thrift.Response{}, err
+		return thrift.Response{}, yarpcerrors.InvalidArgumentErrorf(
+			"could not decode Thrift request for service 'Hello' procedure 'Echo': %w", err)
 	}
 
 	success, err := h.impl.Echo(ctx, args.Echo)