+++ internal/inboundcall/metadata_test.go
@@ -18,23 +18,19 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package yarpctest
+package inboundcall
 
 import (
-	"net"
-	"regexp"
+	"context"
 	"testing"
-
-	"github.com/stretchr/testify/require"
 )
 
-var regex = regexp.MustCompile(`127\.0\.0\.1:[0-9]+`)
-
-func TestZeroAddrToHostPort(t *testing.T) {
-	for i := 0; i < 10; i++ {
-		listener, err := net.Listen("tcp", "127.0.0.1:0")
-		require.NoError(t, err)
-		require.True(t, regex.MatchString(ZeroAddrToHostPort(listener.Addr())))
-		require.NoError(t, listener.Close())
+func TestRoundTrip(t *testing.T) {
+	give := md{}
+	got, ok := GetMetadata(WithMetadata(context.Background(), give))
+	if !ok || got != give {
+		t.Errorf("could not round trip metadata")
 	}
 }
+
+type md struct{ Metadata }