+++ transport/grpc/handler_test.go
@@ -25,8 +25,12 @@ import (
 	"net"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"go.uber.org/yarpc/yarpcerrors"
+	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
+	"google.golang.org/grpc/status"
 )
 
 func TestInvalidStreamContext(t *testing.T) {
@@ -113,4 +117,26 @@ func TestInvalidStreamMultipleHeaders(t *testing.T) {
 
 	require.Contains(t, err.Error(), "code:invalid-argument")
 	require.Contains(t, err.Error(), "header has more than one value: rpc-caller")
+}
+
+func TestToGRPCError(t *testing.T) {
+	t.Run("nil", func(t *testing.T) {
+		assert.Nil(t, toGRPCError(nil))
+	})
+
+	t.Run("gRPC status", func(t *testing.T) {
+		grpcSt := status.New(codes.InvalidArgument, "foo").Err()
+		assert.Equal(t, grpcSt, toGRPCError(grpcSt), "expected same error given")
+	})
+
+	t.Run("yarpcerror", func(t *testing.T) {
+		msg := "foo"
+		yErr := yarpcerrors.FailedPreconditionErrorf(msg)
+
+		grpcSt, ok := status.FromError(toGRPCError(yErr))
+		require.True(t, ok, "expected gRPC error")
+
+		assert.Equal(t, codes.FailedPrecondition, grpcSt.Code(), "code")
+		assert.Equal(t, msg, grpcSt.Message(), "message mismatch")
+	})
 }