+++ transport/grpc/integration_test.go
@@ -27,10 +27,10 @@ import (
 	"crypto/ecdsa"
 	"crypto/elliptic"
 	"crypto/rand"
+	"crypto/tls"
 	"crypto/x509"
 	"crypto/x509/pkix"
 	"errors"
-	"fmt"
 	"io"
 	"math"
 	"math/big"
@@ -60,6 +60,7 @@ import (
 	"go.uber.org/zap/zaptest"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/status"
 )
 
@@ -91,15 +92,6 @@ func TestGRPCBasic(t *testing.T) {
 		assert.NoError(t, err)
 		assert.Equal(t, "bar", value)
 	})
-}
-
-func expectErrorContains(t *testing.T, err error, contains string) {
-	if contains == "" {
-		assert.NoError(t, err)
-	} else {
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), contains)
-	}
 }
 
 func TestYARPCWellKnownError(t *testing.T) {
@@ -396,6 +388,82 @@ func TestGRPCCompression(t *testing.T) {
 			assert.Equal(t, newMetrics(tt.wantMetrics, map[string]string{
 				"compressor": compressor,
 			}), _metrics)
+		})
+	}
+}
+
+func TestTLSWithYARPCAndGRPC(t *testing.T) {
+	tests := []struct {
+		name           string
+		clientValidity time.Duration
+		serverValidity time.Duration
+		wantErr        bool
+	}{
+		{
+			name:           "valid certs both sides",
+			clientValidity: time.Minute,
+			serverValidity: time.Minute,
+		},
+		{
+			name:           "invalid server cert",
+			clientValidity: time.Minute,
+			serverValidity: -1,
+			wantErr:        true,
+		},
+		{
+			name:           "invalid client cert",
+			clientValidity: -1,
+			serverValidity: time.Minute,
+			wantErr:        true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			scenario := createTLSScenario(t, tt.clientValidity, tt.serverValidity)
+
+			serverCreds := credentials.NewTLS(&tls.Config{
+				GetCertificate: func(_ *tls.ClientHelloInfo) (*tls.Certificate, error) {
+					return &tls.Certificate{
+						Certificate: [][]byte{scenario.ServerCert.Raw},
+						Leaf:        scenario.ServerCert,
+						PrivateKey:  scenario.ServerKey,
+					}, nil
+				},
+				ClientAuth: tls.RequireAndVerifyClientCert,
+				ClientCAs:  scenario.CAs,
+			})
+
+			clientCreds := credentials.NewTLS(&tls.Config{
+				GetClientCertificate: func(_ *tls.CertificateRequestInfo) (*tls.Certificate, error) {
+					return &tls.Certificate{
+						Certificate: [][]byte{scenario.ClientCert.Raw},
+						Leaf:        scenario.ClientCert,
+						PrivateKey:  scenario.ClientKey,
+					}, nil
+				},
+				RootCAs: scenario.CAs,
+			})
+
+			te := testEnvOptions{
+				InboundOptions: []InboundOption{InboundCredentials(serverCreds)},
+				DialOptions:    []DialOption{DialerCredentials(clientCreds)},
+			}
+			te.do(t, func(t *testing.T, e *testEnv) {
+				err := e.SetValueYARPC(context.Background(), "foo", "bar")
+				if tt.wantErr {
+					assert.Error(t, err)
+				} else {
+					assert.NoError(t, err)
+				}
+
+				err = e.SetValueGRPC(context.Background(), "foo", "bar")
+				if tt.wantErr {
+					assert.Error(t, err)
+				} else {
+					assert.NoError(t, err)
+				}
+			})
 		})
 	}
 }
@@ -757,7 +825,7 @@ func (r *testRouter) Choose(_ context.Context, request *transport.Request) (tran
 			return procedure.HandlerSpec, nil
 		}
 	}
-	return transport.HandlerSpec{}, fmt.Errorf("no procedure for name %s", request.Procedure)
+	return transport.HandlerSpec{}, yarpcerrors.UnimplementedErrorf("no procedure for name %s", request.Procedure)
 }
 
 type tlsScenario struct {
@@ -845,4 +913,44 @@ func createTLSScenario(t *testing.T, clientValidity time.Duration, serverValidit
 		ClientCert: clientCert,
 		ClientKey:  clientKey,
 	}
+}
+
+func TestYARPCErrorsConverted(t *testing.T) {
+	// Ensures that all returned errors are gRPC errors and not YARPC errors
+
+	trans := NewTransport()
+
+	listener, err := net.Listen("tcp", "127.0.0.1:0")
+	require.NoError(t, err)
+	inbound := trans.NewInbound(listener)
+
+	outbound := trans.NewSingleOutbound(listener.Addr().String())
+
+	router := &testRouter{}
+	inbound.SetRouter(router)
+
+	require.NoError(t, trans.Start())
+	defer func() { assert.NoError(t, trans.Stop()) }()
+
+	require.NoError(t, inbound.Start())
+	defer func() { assert.NoError(t, inbound.Stop()) }()
+
+	require.NoError(t, outbound.Start())
+	defer func() { assert.NoError(t, outbound.Stop()) }()
+
+	t.Run("no procedure", func(t *testing.T) {
+		ctx, cancel := context.WithTimeout(context.Background(), time.Second)
+		defer cancel()
+
+		_, err := outbound.Call(ctx, &transport.Request{
+			Caller:    "caller",
+			Service:   "service",
+			Encoding:  "encoding",
+			Procedure: "no procedure",
+			Body:      bytes.NewBufferString("foo-body"),
+		})
+
+		require.Error(t, err)
+		assert.True(t, yarpcerrors.IsUnimplemented(err))
+	})
 }