+++ CHANGELOG.md
@@ -4,6 +4,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.45.0] - 2020-04-21
+### Added
+- gRPC inbound supports introspection, suitable for debug pages.
+- yarpctest: Add `ContextWithCall` function to ease testing of functions that
+  use `yarpc.CallFromContext`.
+- `yarpcerrors` are aligned with the `errors` API introduced in Go 1.13
+  (https://blog.golang.org/go1.13-errors). `yarpcerrors.IsStatus`,
+  `yarpcerrors.FromError`, support wrapped errors, and `yarpcerrors.Status`
+  implements `Unwrap() error`.
+### Changed
+- Inbound Thrift requests that fail during `wire.Value#FromWire` are returned as
+  `yarpcerrors.CodeInvalidArgument`, to indicate a client error.
+- Dropped library dependency on development tools.
+### Fixed
+- gRPC inbounds correctly convert all YARPC error codes to gRPC error codes,
+  outside of handler errors. Previously, well-defined YARPC errors were wrapped
+  with an `Unknown` gRPC code for unimplemented procedures.
+- Fixes `idleConnTimeout` not propgated to the underlying http transport.
+- HTTP does not hold onto connections forever by default. `IdleConnTimeout`
+  now defaults to 15 minutes.
+- `protobuf.GetErrorDetails` can extract error details from wrapped errors.
+- observability: errors returned from Protobuf handlers are logged with their
+  error message instead of `application_error`.
+- `encoding/protobuf` error details (`yarpcerrors.Status` with details) are
+  logged at application error log level. This matches behavior with Thrift
+  exceptions.
+
 ## [1.44.0] - 2020-02-27
 ### Added
 - The TChannel Transport now supports a custom dialer function option.
@@ -1189,6 +1216,7 @@ This release requires regeneration of ThriftRW code.
 
 - Initial release.
 
+[1.45.0]: https://github.com/yarpc/yarpc-go/compare/v1.44.0...v1.45.0
 [1.44.0]: https://github.com/yarpc/yarpc-go/compare/v1.43.0...v1.44.0
 [1.43.0]: https://github.com/yarpc/yarpc-go/compare/v1.42.1...v1.43.0
 [1.42.1]: https://github.com/yarpc/yarpc-go/compare/v1.42.0...v1.42.1